+++ packages/components/src/components/use-theme.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import { srcbookLight, srcbookDark } from '..//lib/code-theme';
+import { srcbookLight, srcbookDark } from '../lib/code-theme.js';
 
 export type ThemeType = 'light' | 'dark';
 export type CodeThemeType = typeof srcbookLight | typeof srcbookDark;