+++ packages/web/src/clients/websocket/index.ts
@@ -27,6 +27,14 @@ import {
   TsServerDefinitionLocationResponsePayloadSchema,
   TsServerDefinitionLocationRequestPayloadSchema,
   TsServerCompletionEntriesPayloadSchema,
+  FileCreatedPayloadSchema,
+  FileUpdatedPayloadSchema,
+  FileRenamedPayloadSchema,
+  FileDeletedPayloadSchema,
+  FilePayloadSchema,
+  PreviewStatusPayloadSchema,
+  PreviewStartPayloadSchema,
+  PreviewStopPayloadSchema,
 } from '@srcbook/shared';
 import Channel from '@/clients/websocket/channel';
 import WebSocketClient from '@/clients/websocket/client';
@@ -80,5 +88,35 @@ export class SessionChannel extends Channel<
       incoming: IncomingSessionEvents,
       outgoing: OutgoingSessionEvents,
     });
+  }
+}
+
+const IncomingAppEvents = {
+  file: FilePayloadSchema,
+  'preview:status': PreviewStatusPayloadSchema,
+};
+
+const OutgoingAppEvents = {
+  'file:created': FileCreatedPayloadSchema,
+  'file:updated': FileUpdatedPayloadSchema,
+  'file:renamed': FileRenamedPayloadSchema,
+  'file:deleted': FileDeletedPayloadSchema,
+  'preview:start': PreviewStartPayloadSchema,
+  'preview:stop': PreviewStopPayloadSchema,
+};
+
+export class AppChannel extends Channel<typeof IncomingAppEvents, typeof OutgoingAppEvents> {
+  appId: string;
+
+  static create(appId: string) {
+    return new AppChannel(appId);
+  }
+
+  constructor(appId: string) {
+    super(client, `app:${appId}`, {
+      incoming: IncomingAppEvents,
+      outgoing: OutgoingAppEvents,
+    });
+    this.appId = appId;
   }
 }