+++ packages/api/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @srcbook/api
 
+## 0.0.9
+
+### Patch Changes
+
+- 459b18d: Deploy all packages
+- Updated dependencies [459b18d]
+  - @srcbook/shared@0.0.9
+
+## 0.0.8
+
+### Patch Changes
+
+- 24c841e: Update websocket client to pass context and connection
+- a063c46: Fix uncaught exceptions in TsServer event handlers after srcbook deletion
+- Updated dependencies [24c841e]
+  - @srcbook/shared@0.0.8
+
 ## 0.0.7
 
 ### Patch Changes