+++ srcbook/CHANGELOG.md
@@ -1,5 +1,23 @@
 # srcbook
 
+## 0.0.9
+
+### Patch Changes
+
+- 459b18d: Deploy all packages
+- Updated dependencies [459b18d]
+  - @srcbook/api@0.0.9
+  - @srcbook/shared@0.0.9
+
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies [24c841e]
+- Updated dependencies [a063c46]
+  - @srcbook/shared@0.0.8
+  - @srcbook/api@0.0.8
+
 ## 0.0.7
 
 ### Patch Changes